backend/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-dotenv==1.2.1
psycopg2-binary==2.9.11
pgvector==0.4.2
requests==2.32.5
cohere==5.13.0
python-jose[cryptography]==3.4.0
pydantic==2.12.5
slowapi==0.1.9
